requirements.txt
aiohttp
discord.py
tabulate